game summary calculation completed

+++ API/Modal/GameSummary.cs
@@ -6,6 +6,6 @@
         public int CurrentRound { get; set; }
         public int WonCount { get; set; }
         public int LostCount { get; set; }
-        public int WinningPercentage { get; set; }
+        public decimal WinningPercentage { get; set; }
     }
 }
